Fixed formatting and indentation in math.h functions, corrected softfloat linker.mk, fixed lgammaf softlock bug, fixed erfcf/erfcl for negative inputs, and fixed the zilog copyright headers so I can sleep peacefully

+++ src/libc/log1pf.c
@@ -1,7 +1,7 @@
 #include <math.h>
 
 float log1pf(float x) {
-    if (fabs(x) <= 0.125) {
+    if (fabsf(x) <= 0.125) {
         // pade(series(ln(1+x),x=0,6,polynom),x,5,3)
         // (-57*x**2-90*x)/(x**3-21*x**2-102*x-90)
         // relative error less than 1e-7
